Add necessary documentation for classes. Move all unresolved classes to unresolved.scala
