+++ zion/src/main/scala/edu/uci/ics/cloudberry/zion/model/impl/JSONParser.scala
@@ -47,13 +47,9 @@ class JSONParser extends IJSONParser {
 object JSONParser {
   //Warn: the order of implicit values matters. The dependence should be initialized earlier
 
-  val resolver = new QueryResolver
-
-  val validator = new QueryValidator
-
   def resolve(query: UnresolvedQuery, schemaMap: Map[String, Schema]): Query = {
-    val resolved = resolver.resolve(query, schemaMap).asInstanceOf[Query]
-    validator.validate(resolved)
+    val resolved = QueryResolver.resolve(query, schemaMap).asInstanceOf[Query]
+    QueryValidator.validate(resolved)
     resolved
   }
 
@@ -264,130 +260,4 @@ object JSONParser {
       )
     ) (UnresolvedQuery.apply, unlift(UnresolvedQuery.unapply))
 
-
-  def toUnresolved(query: Query): UnresolvedQuery =
-    UnresolvedQuery(
-      query.dataset,
-      query.lookup.map(toUnresolved(_)),
-      query.filter.map(toUnresolved(_)),
-      query.unnest.map(toUnresolved(_)),
-      query.groups.map(toUnresolved(_)),
-      query.select.map(toUnresolved(_)),
-      query.globalAggr.map(toUnresolved(_)),
-      query.isEstimable
-    )
-
-  def toUnresolved(lookup: LookupStatement): UnresolvedLookupStatement =
-    UnresolvedLookupStatement(
-      lookup.sourceKeys.map(_.name),
-      lookup.dataset,
-      lookup.lookupKeys.map(_.name),
-      lookup.selectValues.map(_.name),
-      lookup.as.map(_.name)
-    )
-
-  def toUnresolved(aggr: AggregateStatement): UnresolvedAggregateStatement =
-    UnresolvedAggregateStatement(
-      aggr.field.name,
-      aggr.func,
-      aggr.as.name
-    )
-
-  def toUnresolved(filter: FilterStatement): UnresolvedFilterStatement =
-    UnresolvedFilterStatement(
-      filter.field.name,
-      filter.funcOpt,
-      filter.relation,
-      filter.values
-    )
-
-  def toUnresolved(unnest: UnnestStatement): UnresolvedUnnestStatement =
-    UnresolvedUnnestStatement(
-      unnest.field.name,
-      unnest.as.name
-    )
-
-  def toUnresolved(by: ByStatement): UnresolvedByStatement =
-    UnresolvedByStatement(
-      by.field.name,
-      by.funcOpt,
-      by.as.map(_.name)
-    )
-
-  def toUnresolved(group: GroupStatement): UnresolvedGroupStatement =
-    UnresolvedGroupStatement(
-      group.bys.map(toUnresolved(_)),
-      group.aggregates.map(toUnresolved(_))
-    )
-
-
-  def toUnresolved(globalAggr: GlobalAggregateStatement): UnresolvedGlobalAggregateStatement =
-    UnresolvedGlobalAggregateStatement(
-      toUnresolved(globalAggr.aggregate)
-    )
-
-  def toUnresolved(select: SelectStatement): UnresolvedSelectStatement =
-    UnresolvedSelectStatement(
-      select.orderOn.zip(select.order).map {
-        case (field, order) =>
-          order match {
-            case SortOrder.DSC => "-" + field.name
-            case SortOrder.ASC => field.name
-          }
-      },
-      select.limit,
-      select.offset,
-      select.fields.map(_.name)
-
-    )
-
 }
-
-
-case class UnresolvedQuery(dataset: String,
-                           lookup: Seq[UnresolvedLookupStatement] = Seq.empty,
-                           filter: Seq[UnresolvedFilterStatement] = Seq.empty,
-                           unnest: Seq[UnresolvedUnnestStatement] = Seq.empty,
-                           groups: Option[UnresolvedGroupStatement] = None,
-                           select: Option[UnresolvedSelectStatement] = None,
-                           globalAggr: Option[UnresolvedGlobalAggregateStatement] = None,
-                           estimable: Boolean = false
-                          ) extends IReadQuery
-
-case class UnresolvedLookupStatement(sourceKeys: Seq[String],
-                                     dataset: String,
-                                     lookupKeys: Seq[String],
-                                     selectValues: Seq[String],
-                                     as: Seq[String]
-                                    ) extends Statement
-
-case class UnresolvedFilterStatement(field: String,
-                                     funcOpt: Option[TransformFunc],
-                                     relation: Relation,
-                                     values: Seq[Any]
-                                    ) extends Statement
-
-case class UnresolvedUnnestStatement(field: String, as: String)
-
-case class UnresolvedByStatement(field: String,
-                                 funcOpt: Option[GroupFunc],
-                                 as: Option[String]
-                                ) extends Statement
-
-case class UnresolvedAggregateStatement(field: String,
-                                        func: AggregateFunc,
-                                        as: String
-                                       ) extends Statement
-
-case class UnresolvedGroupStatement(bys: Seq[UnresolvedByStatement],
-                                    aggregates: Seq[UnresolvedAggregateStatement]
-                                   ) extends Statement
-
-case class UnresolvedGlobalAggregateStatement(aggregate: UnresolvedAggregateStatement
-                                             ) extends Statement
-
-case class UnresolvedSelectStatement(orderOn: Seq[String],
-                                     limit: Int,
-                                     offset: Int,
-                                     fields: Seq[String]
-                                    ) extends Statement
